api_server/requirements.txt
flask==2.3.3
flask-cors==4.0.0
torch>=1.9.0
torchvision>=0.10.0
pillow>=8.3.0
numpy>=1.21.0
opencv-python>=4.5.0